# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Spector Cognitive Memory & Search API
  description: "Ultra-fast cognitive memory, vector search, and agent orchestration\
    \ platform with biological memory tiers (Working, Episodic, Semantic, Procedural)."
  contact:
    name: Spectrayan Maintainers
    url: https://github.com/spectrayan/spector
    email: support@spectrayan.com
  license:
    name: "Apache License, Version 2.0"
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
- url: /
  description: Current Server Instance
security:
- ApiKeyAuth: []
- BearerAuth: []
...
paths:
...
  /api/v1/credentials/{name}:
    get:
      tags:
      - credential-controller
      operationId: getCredential
      parameters:
      - name: X-Tenant-ID
        in: header
        required: false
        schema:
          type: string
          default: default
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CredentialResponse"
    put:
      tags:
      - credential-controller
      operationId: updateCredential
      parameters:
      - name: X-Tenant-ID
        in: header
        required: false
        schema:
          type: string
          default: default
      - name: name
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCredentialRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CredentialResponse"
    delete:
      tags:
      - credential-controller
      operationId: deleteCredential
      parameters:
      - name: X-Tenant-ID
        in: header
        required: false
        schema:
          type: string
          default: default
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
...
  /api/v1/credentials:
    get:
      tags:
      - credential-controller
      operationId: listCredentials
      parameters:
      - name: X-Tenant-ID
        in: header
        required: false
        schema:
          type: string
          default: default
      - name: userId
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CredentialResponse"
    post:
      tags:
      - credential-controller
      operationId: createCredential
      parameters:
      - name: X-Tenant-ID
        in: header
        required: false
        schema:
          type: string
          default: default
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCredentialRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CredentialResponse"
  /api/v1/credentials/{name}/test:
    post:
      tags:
      - credential-controller
      operationId: testCredential
      parameters:
      - name: X-Tenant-ID
        in: header
        required: false
        schema:
          type: string
          default: default
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
  /api/v1/connectors/test:
    post:
      tags:
      - connector-controller
      operationId: testConnectionConfig
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteConfig"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"
  /api/v1/connectors/routes:
    get:
      tags:
      - connector-controller
      operationId: listRoutes
      parameters:
      - name: tenantId
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteConfig"
    post:
      tags:
      - connector-controller
      operationId: createRoute
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteConfig"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RouteConfig"
  /api/v1/connectors/routes/{id}/test:
    post:
      tags:
      - connector-controller
      operationId: testConnectionForRoute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ConnectionTestResult"
  /api/v1/connectors/routes/{id}/stop:
    post:
      tags:
      - connector-controller
      operationId: stopRoute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RouteConfig"
  /api/v1/connectors/routes/{id}/start:
    post:
      tags:
      - connector-controller
      operationId: startRoute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RouteConfig"
...
  /api/v1/connectors/templates/{id}:
    get:
      tags:
      - connector-controller
      operationId: getTemplate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TemplateDescriptor"
  /api/v1/connectors/routes/{id}:
    get:
      tags:
      - connector-controller
      operationId: getRoute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RouteConfig"
    delete:
      tags:
      - connector-controller
      operationId: deleteRoute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/v1/connectors/routes/{id}/status:
    get:
      tags:
      - connector-controller
      operationId: getRouteStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties: {}
  /api/v1/connectors/routes/{id}/history:
    get:
      tags:
      - connector-controller
      operationId: getExecutionHistory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExecutionRecord"
...
components:
  schemas:
...
    UpdateCredentialRequest:
      type: object
      properties:
        category:
          type: string
          enum:
          - LLM
          - EMBEDDING
          - CHANNEL
          - DATABASE
          - BROKER
          - STORAGE
          - APP
          - VAULT
        provider:
          type: string
        credentialType:
          type: string
          enum:
          - API_KEY
          - BEARER_TOKEN
          - BASIC_AUTH
          - CONNECTION_STRING
          - OAUTH2
          - SERVICE_ACCOUNT_JSON
          - KEY_PAIR
          - APP_ROLE
        secret:
          type: string
        properties:
          type: object
          additionalProperties: {}
        isDefault:
          type: boolean
        description:
          type: string
        expiresAt:
          type: string
          format: date-time
    CredentialResponse:
      type: object
      properties:
        credentialId:
          type: string
        tenantId:
          type: string
        userId:
          type: string
        name:
          type: string
        category:
          type: string
          enum:
          - LLM
          - EMBEDDING
          - CHANNEL
          - DATABASE
          - BROKER
          - STORAGE
          - APP
          - VAULT
        provider:
          type: string
        credentialType:
          type: string
          enum:
          - API_KEY
          - BEARER_TOKEN
          - BASIC_AUTH
          - CONNECTION_STRING
          - OAUTH2
          - SERVICE_ACCOUNT_JSON
          - KEY_PAIR
          - APP_ROLE
        maskedPreview:
          type: string
        properties:
          type: object
          additionalProperties: {}
        isDefault:
          type: boolean
        description:
          type: string
        version:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
...
    CreateCredentialRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 128
          minLength: 0
        category:
          type: string
          enum:
          - LLM
          - EMBEDDING
          - CHANNEL
          - DATABASE
          - BROKER
          - STORAGE
          - APP
          - VAULT
        provider:
          type: string
          maxLength: 64
          minLength: 0
        credentialType:
          type: string
          enum:
          - API_KEY
          - BEARER_TOKEN
          - BASIC_AUTH
          - CONNECTION_STRING
          - OAUTH2
          - SERVICE_ACCOUNT_JSON
          - KEY_PAIR
          - APP_ROLE
        secret:
          type: string
          minLength: 1
        properties:
          type: object
          additionalProperties: {}
        isDefault:
          type: boolean
        description:
          type: string
        expiresAt:
          type: string
          format: date-time
      required:
      - category
      - name
      - provider
      - secret
    RouteConfig:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        templateId:
          type: string
        connectorType:
          type: string
        tenantId:
          type: string
        source:
          type: string
        schedule:
          type: string
        properties:
          type: object
          additionalProperties:
            type: string
        credentialRef:
          type: string
        routeYaml:
          type: string
        status:
          type: string
          enum:
          - DRAFT
          - ACTIVE
          - DISABLED
          - ERROR
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ConnectionTestResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        latencyMs:
          type: integer
          format: int64
...
    ParameterDefinition:
      type: object
      properties:
        name:
          type: string
        displayName:
          type: string
        description:
          type: string
        type:
          type: string
        required:
          type: boolean
        defaultValue:
          type: string
        placeholder:
          type: string
    TemplateDescriptor:
      type: object
      properties:
        templateId:
          type: string
        displayName:
          type: string
        description:
          type: string
        icon:
          type: string
        category:
          type: string
        connectorType:
          type: string
        parameters:
          type: array
          items:
            $ref: "#/components/schemas/ParameterDefinition"
        requiresCredential:
          type: boolean
        builtIn:
          type: boolean
        routeYaml:
          type: string
        createdAt:
          type: string
          format: date-time
    ExecutionRecord:
      type: object
      properties:
        traceId:
          type: string
        routeId:
          type: string
        tenantId:
          type: string
        status:
          type: string
          enum:
          - RUNNING
          - COMPLETED
          - FAILED
          - CANCELLED
        documentsProcessed:
          type: integer
          format: int32
        errors:
          type: integer
          format: int32
        duration:
          type: string
        startedAt:
          type: string
          format: date-time
        errorMessage:
          type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      description: Spector API Key Authentication
      name: X-API-Key
      in: header
    BearerAuth:
      type: http
      description: OAuth2 / OIDC JWT Bearer Token
      scheme: bearer
      bearerFormat: JWT